re-organize file structure of network service
