+++ packages/syft/src/syft/service/network/network_service.py
@@ -17,11 +17,8 @@ from ...node.credentials import SyftVerifyKey
 from ...node.worker_settings import WorkerSettings
 from ...serde.serializable import serializable
 from ...service.settings.settings import NodeSettings
-from ...store.document_store import BaseUIDStoreStash
 from ...store.document_store import DocumentStore
 from ...store.document_store import PartitionKey
-from ...store.document_store import PartitionSettings
-from ...store.document_store import QueryKeys
 from ...types.grid_url import GridURL
 from ...types.transforms import TransformContext
 from ...types.transforms import keep
@@ -32,7 +29,6 @@ from ...types.uid import UID
 from ...util.telemetry import instrument
 from ...util.util import prompt_warning_message
 from ..context import AuthedServiceContext
-from ..data_subject.data_subject import NamePartitionKey
 from ..metadata.node_metadata import NodeMetadata
 from ..request.request import Request
 from ..request.request import RequestStatus
@@ -49,6 +45,7 @@ from ..user.user_roles import DATA_OWNER_ROLE_LEVEL
 from ..user.user_roles import GUEST_ROLE_LEVEL
 from ..warnings import CRUDWarning
 from .association_request import AssociationRequestChange
+from .network_stash import NetworkStash
 from .node_peer import NodePeer
 from .node_peer import NodePeerUpdate
 from .routes import HTTPNodeRoute
@@ -68,81 +65,6 @@ class NodePeerAssociationStatus(Enum):
     PEER_ASSOCIATED = "PEER_ASSOCIATED"
     PEER_ASSOCIATION_PENDING = "PEER_ASSOCIATION_PENDING"
     PEER_NOT_FOUND = "PEER_NOT_FOUND"
-
-
-@instrument
-@serializable()
-class NetworkStash(BaseUIDStoreStash):
-    object_type = NodePeer
-    settings: PartitionSettings = PartitionSettings(
-        name=NodePeer.__canonical_name__, object_type=NodePeer
-    )
-
-    def __init__(self, store: DocumentStore) -> None:
-        super().__init__(store=store)
-
-    def get_by_name(
-        self, credentials: SyftVerifyKey, name: str
-    ) -> Result[NodePeer | None, str]:
-        qks = QueryKeys(qks=[NamePartitionKey.with_obj(name)])
-        return self.query_one(credentials=credentials, qks=qks)
-
-    def update(
-        self,
-        credentials: SyftVerifyKey,
-        peer_update: NodePeerUpdate,
-        has_permission: bool = False,
-    ) -> Result[NodePeer, str]:
-        valid = self.check_type(peer_update, NodePeerUpdate)
-        if valid.is_err():
-            return SyftError(message=valid.err())
-        return super().update(credentials, peer_update, has_permission=has_permission)
-
-    def create_or_update_peer(
-        self, credentials: SyftVerifyKey, peer: NodePeer
-    ) -> Result[NodePeer, str]:
-        """
-        Update the selected peer and its route priorities if the peer already exists
-        If the peer does not exist, simply adds it to the database.
-
-        Args:
-            credentials (SyftVerifyKey): The credentials used to authenticate the request.
-            peer (NodePeer): The peer to be updated or added.
-
-        Returns:
-            Result[NodePeer, str]: The updated or added peer if the operation
-            was successful, or an error message if the operation failed.
-        """
-        valid = self.check_type(peer, NodePeer)
-        if valid.is_err():
-            return SyftError(message=valid.err())
-
-        existing = self.get_by_uid(credentials=credentials, uid=peer.id)
-        if existing.is_ok() and existing.ok() is not None:
-            existing_peer: NodePeer = existing.ok()
-            existing_peer.update_routes(peer.node_routes)
-            peer_update = NodePeerUpdate(
-                id=peer.id, node_routes=existing_peer.node_routes
-            )
-            result = self.update(credentials, peer_update)
-            return result
-        else:
-            result = self.set(credentials, peer)
-            return result
-
-    def get_by_verify_key(
-        self, credentials: SyftVerifyKey, verify_key: SyftVerifyKey
-    ) -> Result[NodePeer | None, SyftError]:
-        qks = QueryKeys(qks=[VerifyKeyPartitionKey.with_obj(verify_key)])
-        return self.query_one(credentials, qks)
-
-    def get_by_node_type(
-        self, credentials: SyftVerifyKey, node_type: NodeType
-    ) -> Result[list[NodePeer], SyftError]:
-        qks = QueryKeys(qks=[NodeTypePartitionKey.with_obj(node_type)])
-        return self.query_all(
-            credentials=credentials, qks=qks, order_by=OrderByNamePartitionKey
-        )
 
 
 @instrument
